requirements.txt
azure-keyvault-secrets
azure-identity
snowflake-connector-python
